refactor(moodle): split 100 direct tools into 7 haiku-skills for context efficiency
Replace the monolithic agent (100 @agent.tool_plain decorators, ~15k tokens
of tool schemas per request) with a skill-based router architecture. The
router LLM now sees a single execute_skill tool plus a ~500-token skill
catalog. Each skill spawns an isolated subagent with only its domain tools
loaded (9-19 tools each). Skills: moodle-courses, moodle-users,
moodle-organisation, moodle-certifications, moodle-programs, moodle-rules,
moodle-reporting. 317 tests pass with 100% coverage on agent.py and
skills.py. All 57 E2E scenarios verified against live sandbox.
